requirements.txt
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.2
distlib==0.3.9
Django==5.1.4
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
filelock==3.16.1
idna==3.10
inflection==0.5.1
packaging==25.0
pillow==11.1.0
platformdirs==4.3.6
PyJWT==2.9.0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
setuptools==80.9.0
sqlparse==0.5.3
sslcommerz_python_api==1.0.1
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.4.0
virtualenv==20.28.1
